* Refactor "scenes" directory and classes structure.
