+++ src/controllers/character/CharacterModelsController.ts
@@ -11,6 +11,7 @@ import {
 } from '@babylonjs/core';
 
 import BaseCharacterController from '@mmorpg/controllers/character/BaseCharacterController';
+import PlayerCharacter from '@mmorpg/game-objects/characters/PlayerCharacter';
 import {
 	ICharacterModelsControllerConstructorParams,
 	ICharacterBodyPartsModels,
@@ -18,10 +19,9 @@ import {
 	EMPTY_BODY_PARTS_MODELS_PATHS,
 	EMPTY_BODY_PARTS_MODELS_INSTANCES,
 } from '@mmorpg/interfaces/controllers/character/ICharacterModelsController';
-import PlayerCharacter from '@mmorpg/game-objects/characters/PlayerCharacter';
 import ScenesController from '@mmorpg/controllers/ScenesController';
 import KEY_CODES from '@mmorpg/utils/constants/KEY_CODES';
-import BaseScene from '@mmorpg/scenes/BaseScene';
+import BaseScene from '@mmorpg/scenes/base/BaseScene';
 
 class CharacterModelsController extends BaseCharacterController {
 	private _onMeshLoadedCallback?: () => void;
