+++ src/interfaces/camera/IFollowPlayerCamera.ts
@@ -1,6 +1,6 @@
 import { AbstractMesh } from '@babylonjs/core';
 
-import BaseScene from '@mmorpg/scenes/BaseScene';
+import BaseScene from '@mmorpg/scenes/base/BaseScene';
 
 interface IFollowPlayerCameraConstructorParams {
 	scene: BaseScene;
