+++ src/interfaces/scenes/base/IBaseMapScene.ts
@@ -1,0 +1,5 @@
+import { IBaseSceneConstructorParams } from '@mmorpg/interfaces/scenes/base/IBaseScene';
+
+type IBaseMapSceneConstructorParams = IBaseSceneConstructorParams;
+
+export { IBaseMapSceneConstructorParams };
