+++ src/interfaces/scenes/base/IBaseScene.ts
