+++ src/interfaces/scenes/base/IBaseVirtualScene.ts
@@ -1,0 +1,5 @@
+import { IBaseSceneConstructorParams } from '@mmorpg/interfaces/scenes/base/IBaseScene';
+
+type IBaseVirtualSceneConstructorParams = IBaseSceneConstructorParams;
+
+export { IBaseVirtualSceneConstructorParams };
